Add ads credit promo to all wizard steps.
